+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 * Add support for Python 3.9
+* Add a possibility to append a string to the User-Agent in `B2Http`
 
 ### Removed
 * Drop Python 2 and Python 3.4 support :tada: